+++ src/library.js
@@ -163,6 +163,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     },
     lassoStrokeOpacity: true,
     lassoStrokeSize: true,
+    magnifiedPiles: true,
     orderer: true,
     pileBorderColor: {
       set: value => {
@@ -228,7 +229,6 @@ const createPilingJs = (rootElement, initOptions = {}) => {
       get: 'itemRenderer',
       set: value => [createAction.setItemRenderer(value)]
     },
-    scaledPile: true,
     showGrid: true,
     temporaryDepiledPiles: true,
     tempDepileDirection: true,
@@ -893,7 +893,9 @@ const createPilingJs = (rootElement, initOptions = {}) => {
       isFunction(pileOpacity) ? pileOpacity(pile) : pileOpacity
     );
 
-    pileInstance.scale(isFunction(pileScale) ? pileScale(pile) : pileScale);
+    pileInstance.animateScale(
+      isFunction(pileScale) ? pileScale(pile) : pileScale
+    );
 
     pileInstance.borderSize(
       isFunction(pileBorderSize) ? pileBorderSize(pile) : pileBorderSize
@@ -1584,25 +1586,15 @@ const createPilingJs = (rootElement, initOptions = {}) => {
 
     pileIds.forEach((id, index) => {
       const pile = pileInstances.get(id);
-      const tweener = createTweener({
-        duration: 250,
-        delay: 0,
-        interpolator: interpolateVector,
-        endValue: [centerX, centerY],
-        getter: () => {
-          return [pile.x, pile.y];
-        },
-        setter: xy => {
-          pile.moveTo(...xy, false);
-        },
-        onDone: () => {
-          updatePileBounds(pile.id);
-          if (index === pileIds.length - 1) {
-            store.dispatch(createAction.mergePiles(pileIds, false));
-          }
-        }
-      });
-      animator.add(tweener);
+      const onDone =
+        index === pileIds.length - 1
+          ? () => {
+              updatePileBounds(pile.id);
+              store.dispatch(createAction.mergePiles(pileIds, false));
+            }
+          : identity;
+
+      pile.animateMoveTo(centerX, centerY, { onDone });
     });
   };
 
@@ -1626,7 +1618,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
 
   const scalePile = (pileId, wheelDelta) => {
     const pile = pileInstances.get(pileId);
-    if (pile.scaleByWheel(wheelDelta)) {
+    if (pile.magnifyByWheel(wheelDelta)) {
       updatePileBounds(pileId);
     }
     renderRaf();
@@ -1965,21 +1957,24 @@ const createPilingJs = (rootElement, initOptions = {}) => {
       renderRaf();
     }
 
-    if (state.scaledPiles !== newState.scaledPiles) {
-      state.scaledPiles
+    if (state.magnifiedPiles !== newState.magnifiedPiles) {
+      state.magnifiedPiles
         .map(scaledPile => pileInstances.get(scaledPile))
         .filter(scaledPileInstance => scaledPileInstance)
         .forEach(scaledPileInstance => {
-          scaledPileInstance.scale(1);
+          // We currently allow only one item to be magnified up so all
+          // previously magnified piles are reset
+          scaledPileInstance.unmagnify();
           updatePileBounds(scaledPileInstance.id);
           activePile.removeChildren();
           normalPiles.addChild(scaledPileInstance.graphics);
         });
 
-      newState.scaledPiles
+      newState.magnifiedPiles
         .map(scaledPile => pileInstances.get(scaledPile))
         .filter(scaledPileInstance => scaledPileInstance)
         .forEach(scaledPileInstance => {
+          scaledPileInstance.magnify();
           activePile.addChild(scaledPileInstance.graphics);
         });
 
@@ -2154,7 +2149,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
           pile.items.forEach(item => {
             item.tmpAbsX = pileGfx.x;
             item.tmpAbsY = pileGfx.y;
-            item.tmpRelScale = pile.scale();
+            item.tmpRelScale = pile.scale;
           });
           store.dispatch(
             createAction.mergePiles([pileId, collidePiles[0].id], true)
@@ -2205,7 +2200,22 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     });
   };
 
-  const handleDragStartPile = ({ pileId }) => {
+  const handleDragStartPile = ({ pileId, event }) => {
+    const pile = pileInstances.get(pileId);
+
+    if (pile && pile.isMagnified) {
+      const mousePos = event.data.getLocalPosition(pile.graphics.parent);
+
+      pile.graphics.draggingMouseOffset[0] /= pile.magnification;
+      pile.graphics.draggingMouseOffset[1] /= pile.magnification;
+      pile.animateMoveTo(
+        mousePos.x - pile.graphics.draggingMouseOffset[0],
+        mousePos.y - pile.graphics.draggingMouseOffset[1]
+      );
+    }
+
+    store.dispatch(createAction.setMagnifiedPiles([]));
+
     activePile.addChild(pileInstances.get(pileId).graphics);
     handleHighlightPile(pileId);
   };
@@ -2251,14 +2261,13 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     renderRaf();
   };
 
-  const scaleBtnClick = (contextMenuElement, pileId) => () => {
+  const pileMagnificationHandler = (contextMenuElement, pileId) => () => {
     const pile = pileInstances.get(pileId);
-    if (pile.scale() > 1) {
-      store.dispatch(createAction.setScaledPiles([]));
+    if (pile.isMagnified) {
+      store.dispatch(createAction.setMagnifiedPiles([]));
     } else {
-      store.dispatch(createAction.setScaledPiles([pileId]));
+      store.dispatch(createAction.setMagnifiedPiles([pileId]));
     }
-    pile.scaleToggle();
 
     hideContextMenu(contextMenuElement);
   };
@@ -2343,12 +2352,11 @@ const createPilingJs = (rootElement, initOptions = {}) => {
           results.forEach(result => {
             const pile = pileInstances.get(result.id);
             if (pile.graphics.isHover) {
-              if (pile.scale() > 1) {
-                store.dispatch(createAction.setScaledPiles([]));
+              if (pile.isMagnified) {
+                store.dispatch(createAction.setMagnifiedPiles([]));
               } else {
-                store.dispatch(createAction.setScaledPiles([result.id]));
+                store.dispatch(createAction.setMagnifiedPiles([result.pileId]));
               }
-              pile.scaleToggle();
             }
           });
         } else {
@@ -2361,7 +2369,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
         }
       } else {
         store.dispatch(createAction.setFocusedPiles([]));
-        store.dispatch(createAction.setScaledPiles([]));
+        store.dispatch(createAction.setMagnifiedPiles([]));
       }
     }
   };
@@ -2424,7 +2432,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
 
       if (result.length !== 0) {
         event.preventDefault();
-        store.dispatch(createAction.setScaledPiles([result[0].id]));
+        store.dispatch(createAction.setMagnifiedPiles([result[0].pileId]));
         scalePile(result[0].id, normalizeWheel(event).pixelY);
       }
     } else if (isPanZoom) {
@@ -2530,7 +2538,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     const tempDepileBtn = element.querySelector('#temp-depile-button');
     const toggleGridBtn = element.querySelector('#grid-button');
     const alignBtn = element.querySelector('#align-button');
-    const scaleBtn = element.querySelector('#scale-button');
+    const magnifyBtn = element.querySelector('#magnify-button');
 
     // click on pile
     if (clickedOnPile) {
@@ -2552,13 +2560,13 @@ const createPilingJs = (rootElement, initOptions = {}) => {
       } else if (pile.isTempDepiled) {
         depileBtn.setAttribute('disabled', '');
         depileBtn.setAttribute('class', 'inactive');
-        scaleBtn.setAttribute('disabled', '');
-        scaleBtn.setAttribute('class', 'inactive');
+        magnifyBtn.setAttribute('disabled', '');
+        magnifyBtn.setAttribute('class', 'inactive');
         tempDepileBtn.innerHTML = 'close temp depile';
       }
 
-      if (pile.scale() > 1) {
-        scaleBtn.innerHTML = 'Scale Down';
+      if (pile.isMagnified) {
+        magnifyBtn.innerHTML = 'Unmagnify';
       }
 
       element.style.display = 'block';
@@ -2581,9 +2589,9 @@ const createPilingJs = (rootElement, initOptions = {}) => {
         tempDepileBtnClick(element, pile.id, event),
         false
       );
-      scaleBtn.addEventListener(
+      magnifyBtn.addEventListener(
         'click',
-        scaleBtnClick(element, pile.id),
+        pileMagnificationHandler(element, pile.id),
         false
       );
 
@@ -2604,7 +2612,7 @@ const createPilingJs = (rootElement, initOptions = {}) => {
     } else {
       depileBtn.style.display = 'none';
       tempDepileBtn.style.display = 'none';
-      scaleBtn.style.display = 'none';
+      magnifyBtn.style.display = 'none';
 
       if (showGrid) {
         toggleGridBtn.innerHTML = 'Hide Grid';