+++ src/store.js
@@ -128,7 +128,7 @@ const [pileItemRotation, setPileItemRotation] = setter(
 
 const [focusedPiles, setFocusedPiles] = setter('focusedPiles', []);
 
-const [scaledPiles, setScaledPiles] = setter('scaledPiles', []);
+const [magnifiedPiles, setMagnifiedPiles] = setter('magnifiedPiles', []);
 
 // 'originalPos' and 'closestPos'
 const [depileMethod, setDepileMethod] = setter('depileMethod', 'originalPos');
@@ -406,7 +406,7 @@ const createStore = () => {
     previewSpacing,
     randomOffsetRange,
     randomRotationRange,
-    scaledPiles,
+    magnifiedPiles,
     showGrid,
     tempDepileDirection,
     tempDepileOneDNum,
@@ -473,7 +473,7 @@ export const createAction = {
   setPileItemAlignment,
   setPileItemRotation,
   setFocusedPiles,
-  setScaledPiles,
+  setMagnifiedPiles,
   setShowGrid,
   setDepileMethod,
   setDepiledPile,